moved out remote screenshare to separate file
